Split up measures views and forms code
