+++ common/urls.py
@@ -11,7 +11,7 @@ from django.urls import register_converter
 
 from common.path_converters import NumericSIDConverter
 from common.views import pages
-from measures import views as measure_views
+from measures.views import DutySentenceReference
 
 register_converter(NumericSIDConverter, "sid")
 
@@ -23,7 +23,7 @@ urlpatterns = [
     path("app-info", pages.AppInfoView.as_view(), name="app-info"),
     path(
         "duties/",
-        measure_views.DutySentenceReference.as_view(),
+        DutySentenceReference.as_view(),
         name="duties",
     ),
     path(
