+++ measures/forms/fields.py
@@ -1,0 +1,5 @@
+from crispy_forms_gds.layout import Field
+
+
+class MeasureConditionComponentDuty(Field):
+    template = "components/measure_condition_component_duty/template.jinja"
