+++ measures/tests/test_bulk_processing.py
@@ -80,7 +80,7 @@ def test_cancel_task(
 
 
 @patch("measures.parsers.DutySentenceParser")
-@patch("measures.forms.LarkDutySentenceParser")
+@patch("measures.forms.wizard.LarkDutySentenceParser")
 def test_bulk_creator_get_forms_cleaned_data(
     mock_lark_duty_sentence_parser,
     mock_duty_sentence_parser,
@@ -168,7 +168,7 @@ def test_bulk_creator_get_forms_cleaned_data(
 
 
 @patch("measures.parsers.DutySentenceParser")
-@patch("measures.forms.LarkDutySentenceParser")
+@patch("measures.forms.wizard.LarkDutySentenceParser")
 def test_bulk_creator_get_forms_cleaned_data_errors(
     mock_lark_duty_sentence_parser,
     mock_duty_sentence_parser,
